planner: move physical PhysicalExpand related logic into physical op dir.
