+++ pkg/planner/core/physical_plans.go
@@ -17,7 +17,6 @@ package core
 import (
 	"fmt"
 	"strconv"
-	"strings"
 	"unsafe"
 
 	"github.com/pingcap/errors"
@@ -39,7 +38,6 @@ import (
 	"github.com/pingcap/tidb/pkg/sessionctx/vardef"
 	"github.com/pingcap/tidb/pkg/statistics"
 	"github.com/pingcap/tidb/pkg/types"
-	"github.com/pingcap/tidb/pkg/util/plancodec"
 	"github.com/pingcap/tidb/pkg/util/ranger"
 	"github.com/pingcap/tidb/pkg/util/size"
 	"github.com/pingcap/tidb/pkg/util/stringutil"
@@ -912,90 +910,6 @@ func (p *PhysicalExchangeReceiver) MemoryUsage() (sum int64) {
 	for _, frag := range p.frags {
 		sum += frag.MemoryUsage()
 	}
-	return
-}
-
-// PhysicalExpand is used to expand underlying data sources to feed different grouping sets.
-type PhysicalExpand struct {
-	// data after repeat-OP will generate a new grouping-ID column to indicate what grouping set is it for.
-	physicalop.PhysicalSchemaProducer
-
-	// generated grouping ID column itself.
-	GroupingIDCol *expression.Column
-
-	// GroupingSets is used to define what kind of group layout should the underlying data follow.
-	// For simple case: select count(distinct a), count(distinct b) from t; the grouping expressions are [a] and [b].
-	GroupingSets expression.GroupingSets
-
-	// The level projections is generated from grouping sets，make execution more clearly.
-	LevelExprs [][]expression.Expression
-
-	// The generated column names. Eg: "grouping_id" and so on.
-	ExtraGroupingColNames []string
-}
-
-// Init only assigns type and context.
-func (p PhysicalExpand) Init(ctx base.PlanContext, stats *property.StatsInfo, offset int, props ...*property.PhysicalProperty) *PhysicalExpand {
-	p.BasePhysicalPlan = physicalop.NewBasePhysicalPlan(ctx, plancodec.TypeExpand, &p, offset)
-	p.SetChildrenReqProps(props)
-	p.SetStats(stats)
-	return &p
-}
-
-// Clone implements op.PhysicalPlan interface.
-func (p *PhysicalExpand) Clone(newCtx base.PlanContext) (base.PhysicalPlan, error) {
-	if len(p.LevelExprs) > 0 {
-		return p.cloneV2(newCtx)
-	}
-	np := new(PhysicalExpand)
-	np.SetSCtx(newCtx)
-	base, err := p.PhysicalSchemaProducer.CloneWithSelf(newCtx, np)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	np.PhysicalSchemaProducer = *base
-	// clone ID cols.
-	np.GroupingIDCol = p.GroupingIDCol.Clone().(*expression.Column)
-
-	// clone grouping expressions.
-	clonedGroupingSets := make([]expression.GroupingSet, 0, len(p.GroupingSets))
-	for _, one := range p.GroupingSets {
-		clonedGroupingSets = append(clonedGroupingSets, one.Clone())
-	}
-	np.GroupingSets = clonedGroupingSets
-	return np, nil
-}
-
-func (p *PhysicalExpand) cloneV2(newCtx base.PlanContext) (base.PhysicalPlan, error) {
-	np := new(PhysicalExpand)
-	base, err := p.PhysicalSchemaProducer.CloneWithSelf(newCtx, np)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	np.PhysicalSchemaProducer = *base
-	// clone level projection expressions.
-	for _, oneLevelProjExprs := range p.LevelExprs {
-		np.LevelExprs = append(np.LevelExprs, util.CloneExprs(oneLevelProjExprs))
-	}
-
-	// clone generated column names.
-	for _, name := range p.ExtraGroupingColNames {
-		np.ExtraGroupingColNames = append(np.ExtraGroupingColNames, strings.Clone(name))
-	}
-	return np, nil
-}
-
-// MemoryUsage return the memory usage of PhysicalExpand
-func (p *PhysicalExpand) MemoryUsage() (sum int64) {
-	if p == nil {
-		return
-	}
-
-	sum = p.PhysicalSchemaProducer.MemoryUsage() + size.SizeOfSlice + int64(cap(p.GroupingSets))*size.SizeOfPointer
-	for _, gs := range p.GroupingSets {
-		sum += gs.MemoryUsage()
-	}
-	sum += p.GroupingIDCol.MemoryUsage()
 	return
 }
